+++ CHANGES.rst
@@ -1,3 +1,12 @@
+Changes in Matrix iOS SDK in 0.6.7 (2016-05-04)
+===============================================
+
+Improvements:
+ * Presence: Manage the currently_active parameter.
+ * MXRestClient: Add API to reset the account password.
+ * Ability to report abuse
+ * Ability to ignore users
+
 Changes in Matrix iOS SDK in 0.6.6 (2016-04-26)
 ===============================================